requirements.txt
# ============================================================
# Landslide4Sense — Detección de Deslizamientos con Deep Learning
# requirements.txt — Dependencias Python
# Proyecto Final — Aprendizaje de Máquinas Aplicado a IA 2026
# ============================================================

# ── Core científico ──────────────────────────────────────────
numpy>=1.24.0,<2.0
scipy>=1.11.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
Pillow>=10.0.0

# ── Datos geoespaciales / multiespectrales ────────────────────
h5py>=3.9.0
rasterio>=1.3.8
pyproj>=3.6.0

# ── Deep Learning ────────────────────────────────────────────
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# ── Arquitecturas de segmentación ────────────────────────────
segmentation-models-pytorch>=0.3.3
timm>=0.9.7                     # EfficientNet-B4 y backbones adicionales

# ── Utilidades de entrenamiento ───────────────────────────────
tqdm>=4.66.0
tensorboard>=2.14.0
pyyaml>=6.0.1
einops>=0.7.0

# ── Métricas y evaluación ─────────────────────────────────────
torchmetrics>=1.2.0

# ── Notebooks ────────────────────────────────────────────────
jupyter>=1.0.0
ipywidgets>=8.1.0
ipykernel>=6.25.0

# ── Calidad de código ─────────────────────────────────────────
black>=23.9.1
isort>=5.12.0
flake8>=6.1.0

# ── Opcionales (para entornos con GPU) ────────────────────────
# nvidia-ml-py3>=7.352.0        # GPU monitoring
# onnx>=1.14.0                  # Export a ONNX
# onnxruntime-gpu>=1.16.0       # Inferencia optimizada
